requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-common-sql==1.30.2
airflow-provider-duckdb==0.2.0
airflow-ai-sdk[openai]
sentence-transformers
